Move ext/common to runtime/common
