+++ flashlight/app/asr/tools/alignment/Align.cpp
@@ -14,8 +14,8 @@
 #include "flashlight/pkg/speech/data/Utils.h"
 #include "flashlight/pkg/speech/runtime/runtime.h"
 #include "flashlight/app/asr/tools/alignment/Utils.h"
-#include "flashlight/ext/common/SequentialBuilder.h"
-#include "flashlight/ext/common/Serializer.h"
+#include "flashlight/pkg/runtime/common/SequentialBuilder.h"
+#include "flashlight/pkg/runtime/common/Serializer.h"
 #include "flashlight/fl/flashlight.h"
 #include "flashlight/lib/common/System.h"
 #include "flashlight/lib/text/dictionary/Defines.h"
