+++ flashlight/app/objdet/examples/CocoDetr.cpp
@@ -20,8 +20,8 @@
 #include "flashlight/pkg/vision/models/Detr.h"
 #include "flashlight/pkg/vision/nn/Transformer.h"
 #include "flashlight/pkg/runtime/amp/DynamicScaler.h"
-#include "flashlight/ext/common/DistributedUtils.h"
-#include "flashlight/ext/common/Serializer.h"
+#include "flashlight/pkg/runtime/common/DistributedUtils.h"
+#include "flashlight/pkg/runtime/common/Serializer.h"
 #include "flashlight/pkg/vision/dataset/Transforms.h"
 #include "flashlight/fl/meter/meters.h"
 #include "flashlight/fl/optim/optim.h"
