+++ flashlight/pkg/runtime/CMakeLists.txt
@@ -34,6 +34,7 @@ target_sources(
 
 include(${CMAKE_CURRENT_LIST_DIR}/amp/CMakeLists.txt)
 include(${CMAKE_CURRENT_LIST_DIR}/plugin/CMakeLists.txt)
+include(${CMAKE_CURRENT_LIST_DIR}/common/CMakeLists.txt)
 
 target_link_libraries(
   fl_pkg_runtime
