+++ flashlight/pkg/runtime/test/common/SequentialBuilderTest.cpp
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 
-#include "flashlight/ext/common/SequentialBuilder.h"
+#include "flashlight/pkg/runtime/common/SequentialBuilder.h"
 #include "flashlight/fl/common/Init.h"
 #include "flashlight/lib/common/System.h"
