+++ flashlight/pkg/speech/criterion/Seq2SeqCriterion.h
@@ -14,7 +14,7 @@
 #include "flashlight/pkg/speech/criterion/attention/attention.h"
 #include "flashlight/pkg/speech/criterion/attention/window.h"
 
-#include "flashlight/ext/common/DistributedUtils.h"
+#include "flashlight/pkg/runtime/common/DistributedUtils.h"
 
 namespace fl {
 namespace app {
