+++ flashlight/pkg/speech/criterion/TransformerCriterion.h
@@ -14,7 +14,7 @@
 #include "flashlight/pkg/speech/criterion/attention/attention.h"
 #include "flashlight/pkg/speech/criterion/attention/window.h"
 
-#include "flashlight/ext/common/DistributedUtils.h"
+#include "flashlight/pkg/runtime/common/DistributedUtils.h"
 #include "flashlight/fl/contrib/modules/Transformer.h"
 
 namespace fl {
