+++ flashlight/pkg/speech/criterion/backend/cpu/CriterionUtils.cpp
@@ -7,7 +7,7 @@
 
 #include "flashlight/pkg/speech/criterion/CriterionUtils.h"
 
-#include "flashlight/ext/common/DistributedUtils.h"
+#include "flashlight/pkg/runtime/common/DistributedUtils.h"
 #include "flashlight/lib/sequence/criterion/cpu/CriterionUtils.h"
 #include "flashlight/lib/sequence/criterion/cpu/ViterbiPath.h"
