+++ flashlight/pkg/speech/criterion/backend/cuda/ConnectionistTemporalClassificationCriterion.cpp
@@ -13,7 +13,7 @@
 #include "flashlight/pkg/speech/criterion/ConnectionistTemporalClassificationCriterion.h"
 #include "flashlight/pkg/speech/criterion/CriterionUtils.h"
 
-#include "flashlight/ext/common/DistributedUtils.h"
+#include "flashlight/pkg/runtime/common/DistributedUtils.h"
 #include "flashlight/lib/sequence/criterion/cuda/CriterionUtils.cuh"
 
 using namespace fl::ext;
