+++ flashlight/pkg/speech/decoder/ConvLmModule.cpp
@@ -9,7 +9,7 @@
 
 #include <string>
 
-#include "flashlight/ext/common/DistributedUtils.h"
+#include "flashlight/pkg/runtime/common/DistributedUtils.h"
 
 namespace fl {
 namespace app {
