+++ flashlight/pkg/speech/test/runtime/RuntimeTest.cpp
@@ -13,7 +13,7 @@
 #include <gtest/gtest.h>
 
 #include "flashlight/pkg/speech/runtime/runtime.h"
-#include "flashlight/ext/common/Serializer.h"
+#include "flashlight/pkg/runtime/common/Serializer.h"
 #include "flashlight/fl/flashlight.h"
 #include "flashlight/lib/common/System.h"
